+++ tests/projects/c/linker_scripts/src/main.c
@@ -1,0 +1,8 @@
+#include <stdio.h>
+
+void foo();
+
+int main(int argc, char** argv) {
+    foo();
+    return 0;
+}